+++ resources/lang/en/language.php
@@ -559,6 +559,7 @@ return [
         'download'   => 'Download',
         'remove'     => 'Remove',
         'status'     => 'Status',
+        'version'    => 'Version',
         'installing' => 'Installing',
         'guide'      => 'Read document in <a target=_new href="https://s-cart.org/download/template.html">HERE</a>',
     ],