+++ resources/lang/vi/email.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+
+    'reset_password' => [
+        'title'        => 'Xin chào!',
+        'reset_button' => 'Tạo lại mật khẩu',
+        'text1'        => 'Bạn đang nhận email này vì chúng tôi nhận được yêu cầu thiết lập lại mật khẩu cho tài khoản của bạn.',
+        'text2'        => 'Liên kết mật khẩu này sẽ hết hạn trong vòng 60 phút.<br><br>Nếu bạn không yêu cầu một thiết lập lại mật khẩu, bạn không cần làm gì cả.<br><br>Trân trọng,<br>:site_admin',
+        'text3'        => 'Nếu bạn gặp sự cố cách nhấn vào nút ":reset_button", sao chép và dán URL dưới đây vào trình duyệt web của bạn:',
+    ],
+
+];