+++ resources/views/admin/TemplatesManager.blade.php
@@ -14,6 +14,7 @@
                   <th>{{ trans('language.templates.auth') }}</th>
                   <th>{{ trans('language.templates.email') }}</th>
                   <th>{{ trans('language.templates.website') }}</th>
+                  <th>{{ trans('language.templates.version') }}</th>
                   <th>{{ trans('language.templates.status') }}</th>
                 </tr>
                 </thead>
@@ -24,6 +25,7 @@
                      <td>{{ $template['config']['auth']??'' }}</td>
                      <td>{{ $template['config']['email']??'' }}</td>
                      <td>{{ $template['config']['website']??'' }}</td>
+                     <td>{{ $template['config']['version']??'' }}</td>
                       <td>{!! ($templateCurrent == $key)?'<button title="'.trans('language.templates.active').'"  class="btn">'.trans('language.templates.active').'</button >':'<button  onClick="enableTemplate($(this),\''.$key.'\');" title="'.trans('language.templates.inactive').'" data-loading-text="'.trans('language.templates.installing').'" class="btn btn-primary">'.trans('language.templates.inactive').'</button >' !!}</td>
                     </tr>
                   @endforeach