+++ resources/views/templates/default/shop_home.blade.php
@@ -1,4 +1,4 @@
-@extends($theme.'.shop_layout')
+@extends(SITE_THEME.'.shop_layout')
 
 @section('center')
           <div class="features_items"><!--features_items-->
@@ -14,9 +14,9 @@
                             <a class="btn btn-default add-to-cart" onClick="addToCart('{{ $product_new->id }}','default',$(this))"><i class="fa fa-shopping-cart"></i>{{trans('language.add_to_cart')}}</a>
                           </div>
                       @if ($product_new->price != $product_new->getPrice())
-                      <img src="{{ asset($theme_asset.'/images/home/sale.png') }}" class="new" alt="" />
+                      <img src="{{ asset(SITE_THEME_ASSET.'/images/home/sale.png') }}" class="new" alt="" />
                       @elseif($product_new->type == 1)
-                      <img src="{{ asset($theme_asset.'/images/home/new.png') }}" class="new" alt="" />
+                      <img src="{{ asset(SITE_THEME_ASSET.'/images/home/new.png') }}" class="new" alt="" />
                       @endif
                       </div>
                       <div class="choose">
@@ -49,9 +49,9 @@
                             <a class="btn btn-default add-to-cart" onClick="addToCart('{{ $product_hot->id }}','default',$(this))"><i class="fa fa-shopping-cart"></i>{{trans('language.add_to_cart')}}</a>
                           </div>
                           @if ($product_hot->price != $product_hot->getPrice())
-                          <img src="{{ asset($theme_asset.'/images/home/sale.png') }}" class="new" alt="" />
+                          <img src="{{ asset(SITE_THEME_ASSET.'/images/home/sale.png') }}" class="new" alt="" />
                           @elseif($product_hot->type == 1)
-                          <img src="{{ asset($theme_asset.'/images/home/new.png') }}" class="new" alt="" />
+                          <img src="{{ asset(SITE_THEME_ASSET.'/images/home/new.png') }}" class="new" alt="" />
                           @endif
                       </div>
                       <div class="choose">
@@ -99,9 +99,9 @@
                             <a class="btn btn-default add-to-cart" onClick="addToCart('{{ $product->id }}','default',$(this))"><i class="fa fa-shopping-cart"></i>{{trans('language.add_to_cart')}}</a>
                           </div>
                           @if ($product->price != $product->getPrice())
-                          <img src="{{ asset($theme_asset.'/images/home/sale.png') }}" class="new" alt="" />
+                          <img src="{{ asset(SITE_THEME_ASSET.'/images/home/sale.png') }}" class="new" alt="" />
                           @elseif($product->type == 1)
-                          <img src="{{ asset($theme_asset.'/images/home/new.png') }}" class="new" alt="" />
+                          <img src="{{ asset(SITE_THEME_ASSET.'/images/home/new.png') }}" class="new" alt="" />
                           @endif
 
                         </div>